NewSwapExactAmountInCmd: migrate to osmocli / remove from gamm
