+++ x/swaprouter/client/cli/cli_test.go
@@ -8,10 +8,12 @@ import (
 	"github.com/stretchr/testify/suite"
 
 	"github.com/osmosis-labs/osmosis/osmoutils"
+	"github.com/osmosis-labs/osmosis/osmoutils/osmocli"
 	"github.com/osmosis-labs/osmosis/v13/app"
 	"github.com/osmosis-labs/osmosis/v13/x/swaprouter/client/cli"
 	swaprouterqueryproto "github.com/osmosis-labs/osmosis/v13/x/swaprouter/client/queryproto"
 	swaproutertestutil "github.com/osmosis-labs/osmosis/v13/x/swaprouter/client/testutil"
+	"github.com/osmosis-labs/osmosis/v13/x/swaprouter/types"
 
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/crypto/hd"
@@ -30,6 +32,8 @@ type IntegrationTestSuite struct {
 	cfg     network.Config
 	network *network.Network
 }
+
+var testAddresses = osmoutils.CreateRandomAccounts(3)
 
 func (s *IntegrationTestSuite) SetupSuite() {
 	s.T().Log("setting up integration test suite")
@@ -213,68 +217,20 @@ func (s IntegrationTestSuite) TestNewSwapExactAmountOutCmd() {
 // 	}
 // }
 
-func (s IntegrationTestSuite) TestNewSwapExactAmountInCmd() {
-	val := s.network.Validators[0]
-
-	info, _, err := val.ClientCtx.Keyring.NewMnemonic("NewSwapExactAmountIn",
-		keyring.English, sdk.FullFundraiserPath, keyring.DefaultBIP39Passphrase, hd.Secp256k1)
-	s.Require().NoError(err)
-
-	newAddr := sdk.AccAddress(info.GetPubKey().Address())
-
-	_, err = banktestutil.MsgSendExec(
-		val.ClientCtx,
-		val.Address,
-		newAddr,
-		sdk.NewCoins(sdk.NewInt64Coin(s.cfg.BondDenom, 20000), sdk.NewInt64Coin("node0token", 20000)), fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-		fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-		osmoutils.DefaultFeeString(s.cfg),
-	)
-	s.Require().NoError(err)
-
-	testCases := []struct {
-		name string
-		args []string
-
-		expectErr    bool
-		respType     proto.Message
-		expectedCode uint32
-	}{
-		{
-			"swap exact amount in", // osmosisd tx swaprouter swap-exact-amount-in 10stake 3 --swap-route-pool-ids=1 --swap-route-denoms=node0token --from=validator --keyring-backend=test --chain-id=testing --yes
-			[]string{
-				"10stake", "3",
-				fmt.Sprintf("--%s=%d", cli.FlagSwapRoutePoolIds, 1),
-				fmt.Sprintf("--%s=%s", cli.FlagSwapRouteDenoms, "node0token"),
-				fmt.Sprintf("--%s=%s", flags.FlagFrom, newAddr),
-				// common args
-				fmt.Sprintf("--%s=true", flags.FlagSkipConfirmation),
-				fmt.Sprintf("--%s=%s", flags.FlagBroadcastMode, flags.BroadcastBlock),
-				fmt.Sprintf("--%s=%s", flags.FlagFees, sdk.NewCoins(sdk.NewCoin(sdk.DefaultBondDenom, sdk.NewInt(10))).String()),
+func TestNewSwapExactAmountInCmd(t *testing.T) {
+	desc, _ := cli.NewSwapExactAmountInCmd()
+	tcs := map[string]osmocli.TxCliTestCase[*types.MsgSwapExactAmountIn]{
+		"swap exact amount in": {
+			Cmd: "10stake 3 --swap-route-pool-ids=1 --swap-route-denoms=node0token --from=" + testAddresses[0].String(),
+			ExpectedMsg: &types.MsgSwapExactAmountIn{
+				Sender:            testAddresses[0].String(),
+				Routes:            []types.SwapAmountInRoute{{PoolId: 1, TokenOutDenom: "node0token"}},
+				TokenIn:           sdk.NewInt64Coin("stake", 10),
+				TokenOutMinAmount: sdk.NewIntFromUint64(3),
 			},
-			false, &sdk.TxResponse{}, 0,
 		},
 	}
-
-	for _, tc := range testCases {
-		tc := tc
-
-		s.Run(tc.name, func() {
-			cmd := cli.NewSwapExactAmountInCmd()
-			clientCtx := val.ClientCtx
-
-			out, err := clitestutil.ExecTestCLICmd(clientCtx, cmd, tc.args)
-			if tc.expectErr {
-				s.Require().Error(err)
-			} else {
-				s.Require().NoError(err, out.String())
-				s.Require().NoError(clientCtx.Codec.UnmarshalJSON(out.Bytes(), tc.respType), out.String())
-
-				txResp := tc.respType.(*sdk.TxResponse)
-				s.Require().Equal(tc.expectedCode, txResp.Code, out.String())
-			}
-		})
-	}
+	osmocli.RunTxTestCases(t, desc, tcs)
 }
 
 func (s *IntegrationTestSuite) TestGetCmdNumPools() {
